lovable-ai-workspace-main/python-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
aiohttp==3.9.1
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
Pillow==10.1.0
python-dotenv==1.0.0
pandas
python-docx
pymupdf
xlsxwriter
